+++ src/Catesta/Resources/Vault/src/PSVault.Settings.ps1
@@ -1,0 +1,2 @@
+# specify the minumum required major PowerShell version that the build script should validate
+[int]$script:requiredPSVersion = '5'