+++ src/Catesta/Resources/Vault/src/PSVault.build.ps1
@@ -1,4 +1,4 @@
-<#
+﻿<#
 .SYNOPSIS
     An Invoke-Build Build file.
 .DESCRIPTION
@@ -10,7 +10,6 @@
         - FormattingCheck
         - Test
         - DevCC
-        - CreateHelpStart
         - Build
         - InfraTest
         - Archive
@@ -48,17 +47,19 @@ function Test-ManifestBool ($Path) {
 #Default Build
 $str = @()
 $str = 'Clean', 'ValidateRequirements', 'ImportModuleManifest'
+<%
+If ($PLASTER_PARAM_CodingStyle -eq 'Stroustrup' -or $PLASTER_PARAM_CodingStyle -eq 'OTBS' -or $PLASTER_PARAM_CodingStyle -eq 'Allman') {
+    @'
 $str += 'FormattingCheck'
+'@
+}
+%>
 $str += 'Analyze', 'Test'
-$str += 'CreateHelpStart'
 $str += 'Build', 'InfraTest', 'Archive'
 Add-BuildTask -Name . -Jobs $str
 
 #Local testing build process
 Add-BuildTask TestLocal Clean, ImportModuleManifest, Analyze, Test
-
-#Local help file creation process
-Add-BuildTask HelpLocal Clean, ImportModuleManifest, CreateHelpStart
 
 # Pre-build variables to be used by other portions of the script
 Enter-Build {
@@ -73,14 +74,11 @@ Enter-Build {
     $manifestInfo = Import-PowerShellDataFile -Path $script:ModuleManifestFile
     $script:ModuleVersion = $manifestInfo.ModuleVersion
     $script:ModuleDescription = $manifestInfo.Description
-    $script:FunctionsToExport = $manifestInfo.FunctionsToExport
+    $Script:FunctionsToExport = $manifestInfo.FunctionsToExport
 
     $script:TestsPath = Join-Path -Path $BuildRoot -ChildPath 'Tests'
     $script:UnitTestsPath = Join-Path -Path $script:TestsPath -ChildPath 'Unit'
     $script:InfraTestsPath = Join-Path -Path $script:TestsPath -ChildPath 'Infrastructure'
-
-    # $script:ArtifactsPath = Join-Path -Path $BuildRoot -ChildPath "$script:ModuleName\Artifacts"
-    # $script:ArchivePath = Join-Path -Path $BuildRoot -ChildPath "$script:ModuleName\Archive"
 
     $script:ArtifactsPath = Join-Path -Path $BuildRoot -ChildPath 'Artifacts'
     $script:ArchivePath = Join-Path -Path $BuildRoot -ChildPath 'Archive'
@@ -154,19 +152,14 @@ Add-BuildTask Clean {
 Add-BuildTask Analyze {
 
     $scriptAnalyzerParams = @{
-        # Path    = $script:ModuleSourcePath
+        Path    = $script:ModuleSourcePath
         Setting = "PSScriptAnalyzerSettings.psd1"
-        # Recurse = $true
-        # Verbose = $false
+        Recurse = $true
+        Verbose = $false
     }
-
-    $filesToAnalyze = Get-ChildItem -Path $script:ModuleSourcePath -Exclude "PSVault.Extension*" -Recurse
 
     Write-Build White '      Performing Module ScriptAnalyzer checks...'
-    foreach ($file in $filesToAnalyze) {
-        $scriptAnalyzerResults = Invoke-ScriptAnalyzer @scriptAnalyzerParams -Path $file.FullName
-    }
-    # $scriptAnalyzerResults = Invoke-ScriptAnalyzer @scriptAnalyzerParams
+    $scriptAnalyzerResults = Invoke-ScriptAnalyzer @scriptAnalyzerParams
 
     if ($scriptAnalyzerResults) {
         $scriptAnalyzerResults | Format-Table
@@ -203,16 +196,43 @@ Add-BuildTask AnalyzeTests -After Analyze {
 
 #Synopsis: Analyze scripts to verify if they adhere to desired coding format (Stroustrup / OTBS / Allman)
 Add-BuildTask FormattingCheck {
+
+<%
+    if ($PLASTER_PARAM_CodingStyle -eq 'Stroustrup') {
+        @'
     $scriptAnalyzerParams = @{
         Setting     = 'CodeFormattingStroustrup'
-        ExcludeRule = @(
-            'PSUseConsistentIndentation',
-            'PSUseConsistentWhitespace'
-        )
-        # ExcludeRule = 'PSUseConsistentWhitespace'
+        ExcludeRule = 'PSUseConsistentWhitespace'
         Recurse     = $true
         Verbose     = $false
     }
+'@
+    }
+%>
+<%
+    if ($PLASTER_PARAM_CodingStyle -eq 'OTBS') {
+        @'
+    $scriptAnalyzerParams = @{
+        Setting     = 'CodeFormattingOTBS'
+        ExcludeRule = 'PSUseConsistentWhitespace'
+        Recurse     = $true
+        Verbose     = $false
+    }
+'@
+    }
+%>
+<%
+    if ($PLASTER_PARAM_CodingStyle -eq 'Allman') {
+        @'
+    $scriptAnalyzerParams = @{
+        Setting     = 'CodeFormattingAllman'
+        ExcludeRule = 'PSUseConsistentWhitespace'
+        Recurse     = $true
+        Verbose     = $false
+    }
+'@
+    }
+%>
 
     Write-Build White '      Performing script formatting checks...'
     $scriptAnalyzerResults = Get-ChildItem -Path $script:ModuleSourcePath -Exclude "*.psd1" | Invoke-ScriptAnalyzer @scriptAnalyzerParams
@@ -243,7 +263,7 @@ Add-BuildTask Test {
             PassThru                     = $true
             Verbose                      = $false
             EnableExit                   = $false
-            CodeCoverage                 = "$ModuleName\*\*.ps1"
+            CodeCoverage                 = "$ModuleName\*\*.psm1"
             CodeCoverageOutputFile       = "$codeCovPath\CodeCoverage.xml"
             CodeCoverageOutputFileFormat = 'JaCoCo'
             OutputFile                   = "$testOutPutPath\PesterTests.xml"
@@ -267,8 +287,7 @@ Add-BuildTask Test {
         Assert-Build($numberFails -eq 0) ('Failed "{0}" unit tests.' -f $numberFails)
 
         # Ensure our builds fail until if below a minimum defined code test coverage threshold
-        $coverageThreshold = 95
-        $coveragePercent = '{0:N2}' -f ($testResults.CodeCoverage.NumberOfCommandsExecuted / $testResults.CodeCoverage.NumberOfCommandsAnalyzed * 100)
+        $coverageThreshold = 50
 
         if ($testResults.CodeCoverage.NumberOfCommandsExecuted -ne 0) {
             $coveragePercent = '{0:N2}' -f ($testResults.CodeCoverage.NumberOfCommandsExecuted / $testResults.CodeCoverage.NumberOfCommandsAnalyzed * 100)
@@ -285,7 +304,6 @@ Add-BuildTask Test {
             else {
                 Write-Build Cyan "      $('Covered {0}% of {1} analyzed commands in {2} files.' -f $coveragePercent,$testResults.CodeCoverage.NumberOfCommandsAnalyzed,$testResults.CodeCoverage.NumberOfFilesAnalyzed)"
                 Write-Build Green '      ...Pester Unit Tests Complete!'
-
             }
         }
         else {
@@ -308,133 +326,11 @@ Add-BuildTask DevCC {
     Write-Build Green '      ...Code Coverage report generated!'
 }#DevCC
 
-# Synopsis: Build help for module
-Add-BuildTask CreateHelpStart {
-    Write-Build White '      Performing all help related actions.'
-
-    Write-Build Gray '           Importing platyPS v0.12.0 ...'
-    Import-Module platyPS -RequiredVersion 0.12.0 -ErrorAction Stop
-    Write-Build Gray '           ...platyPS imported successfully.'
-}#CreateHelpStart
-
-# Synopsis: Build markdown help files for module and fail if help information is missing
-Add-BuildTask CreateMarkdownHelp -After CreateHelpStart {
-    $ModulePage = "$($script:ArtifactsPath)\docs\$($ModuleName).md"
-
-    $markdownParams = @{
-        Module         = $ModuleName
-        OutputFolder   = "$($script:ArtifactsPath)\docs\"
-        Force          = $true
-        WithModulePage = $true
-        Locale         = 'en-US'
-        FwLink         = "NA"
-        HelpVersion    = $script:ModuleVersion
-    }
-
-    Write-Build Gray '           Generating markdown files...'
-    $null = New-MarkdownHelp @markdownParams
-    Write-Build Gray '           ...Markdown generation completed.'
-
-    Write-Build Gray '           Replacing markdown elements...'
-    # Replace multi-line EXAMPLES
-    $OutputDir = "$($script:ArtifactsPath)\docs\"
-    $OutputDir | Get-ChildItem -File | ForEach-Object {
-        # fix formatting in multiline examples
-        $content = Get-Content $_.FullName -Raw
-        $newContent = $content -replace '(## EXAMPLE [^`]+?```\r\n[^`\r\n]+?\r\n)(```\r\n\r\n)([^#]+?\r\n)(\r\n)([^#]+)(#)', '$1$3$2$4$5$6'
-        if ($newContent -ne $content) {
-            Set-Content -Path $_.FullName -Value $newContent -Force
-        }
-    }
-    # Replace each missing element we need for a proper generic module page .md file
-    $ModulePageFileContent = Get-Content -Raw $ModulePage
-    $ModulePageFileContent = $ModulePageFileContent -replace '{{Manually Enter Description Here}}', $script:ModuleDescription
-    $Script:FunctionsToExport | ForEach-Object {
-        Write-Build DarkGray "             Updating definition for the following function: $($_)"
-        $TextToReplace = "{{Manually Enter $($_) Description Here}}"
-        $ReplacementText = (Get-Help -Detailed $_).Synopsis
-        $ModulePageFileContent = $ModulePageFileContent -replace $TextToReplace, $ReplacementText
-    }
-
-    $ModulePageFileContent | Out-File $ModulePage -Force -Encoding:utf8
-    Write-Build Gray '           ...Markdown replacements complete.'
-
-    Write-Build Gray '           Verifying GUID...'
-    $MissingGUID = Select-String -Path "$($script:ArtifactsPath)\docs\*.md" -Pattern "(00000000-0000-0000-0000-000000000000)"
-    if ($MissingGUID.Count -gt 0) {
-        Write-Build Yellow '             The documentation that got generated resulted in a generic GUID. Check the GUID entry of your module manifest.'
-        throw 'Missing GUID. Please review and rebuild.'
-    }
-
-    Write-Build Gray '           Checking for missing documentation in md files...'
-    $MissingDocumentation = Select-String -Path "$($script:ArtifactsPath)\docs\*.md" -Pattern "({{.*}})"
-    if ($MissingDocumentation.Count -gt 0) {
-        Write-Build Yellow '             The documentation that got generated resulted in missing sections which should be filled out.'
-        Write-Build Yellow '             Please review the following sections in your comment based help, fill out missing information and rerun this build:'
-        Write-Build Yellow '             (Note: This can happen if the .EXTERNALHELP CBH is defined for a function before running this build.)'
-        Write-Build Yellow "             Path of files with issues: $($script:ArtifactsPath)\docs\"
-        $MissingDocumentation | Select-Object FileName, LineNumber, Line | Format-Table -AutoSize
-        throw 'Missing documentation. Please review and rebuild.'
-    }
-
-    Write-Build Gray '           Checking for missing SYNOPSIS in md files...'
-    $fSynopsisOutput = @()
-    $synopsisEval = Select-String -Path "$($script:ArtifactsPath)\docs\*.md" -Pattern "^## SYNOPSIS$" -Context 0, 1
-    $synopsisEval | ForEach-Object {
-        $chAC = $_.Context.DisplayPostContext.ToCharArray()
-        if ($null -eq $chAC) {
-            $fSynopsisOutput += $_.FileName
-        }
-    }
-    if ($fSynopsisOutput) {
-        Write-Build Yellow "             The following files are missing SYNOPSIS:"
-        $fSynopsisOutput
-        throw 'SYNOPSIS information missing. Please review.'
-    }
-
-    # Write-Host '      Creating markdown documentation with PlatyPS'
-    # Write-Host -ForegroundColor Green '...Complete!'
-    Write-Build Gray '           ...Markdown generation complete.'
-}#CreateMarkdownHelp
-
-# Synopsis: Build the external xml help file from markdown help files with PlatyPS
-Add-BuildTask CreateExternalHelp -After CreateMarkdownHelp {
-    Write-Build Gray '           Creating external xml help file...'
-    $null = New-ExternalHelp "$($script:ArtifactsPath)\docs" -OutputPath "$($script:ArtifactsPath)\en-US\" -Force
-    Write-Build Gray '           ...External xml help file created!'
-}#CreateExternalHelp
-
-Add-BuildTask CreateHelpComplete -After CreateExternalHelp {
-    Write-Build Green '      ...CreateHelp Complete!'
-}#CreateHelpStart
-
-# Synopsis: Replace comment based help (CBH) with external help in all public functions for this project
-Add-BuildTask UpdateCBH -After AssetCopy {
-    $ExternalHelp = @"
-<#
-.EXTERNALHELP $($ModuleName)-help.xml
-#>
-"@
-
-    $CBHPattern = "(?ms)(\<#.*\.SYNOPSIS.*?#>)"
-    Get-ChildItem -Path "$($script:ArtifactsPath)\Public\*.ps1" -File | ForEach-Object {
-        $FormattedOutFile = $_.FullName
-        Write-Output "      Replacing CBH in file: $($FormattedOutFile)"
-        $UpdatedFile = (Get-Content  $FormattedOutFile -raw) -replace $CBHPattern, $ExternalHelp
-        $UpdatedFile | Out-File -FilePath $FormattedOutFile -force -Encoding:utf8
-    }
-}#UpdateCBH
-
 # Synopsis: Copies module assets to Artifacts folder
 Add-BuildTask AssetCopy -Before Build {
     Write-Build Gray '        Copying assets to Artifacts...'
-    Copy-Item -Path "$script:ModuleSourcePath\*" -Destination $script:ArtifactsPath -Exclude *.psd1, *.psm1 -Recurse -ErrorAction Stop
-    Copy-Item -Path "$script:ModuleSourcePath\Resources\Module\src\PSScriptAnalyzerSettings.psd1" -Destination "$script:ArtifactsPath\Resources\Module\src\PSScriptAnalyzerSettings.psd1" -ErrorAction Stop
-    Copy-Item -Path "$script:ModuleSourcePath\Resources\Module\src\Module\Module.psm1" -Destination "$script:ArtifactsPath\Resources\Module\src\Module\Module.psm1" -ErrorAction Stop
-    # Copy-Item -Path "$script:ModuleSourcePath\Resources\Vault\src\PSVault\PSVault.psd1" -Destination "$script:ArtifactsPath\Resources\Vault\src\PSVault\PSVault.psd1" -ErrorAction Stop
-    Copy-Item -Path "$script:ModuleSourcePath\Resources\Vault\src\PSVault\PSVault.Extension\PSVault.Extension.psd1" -Destination "$script:ArtifactsPath\Resources\Vault\src\PSVault\PSVault.Extension\PSVault.Extension.psd1" -ErrorAction Stop
-    Copy-Item -Path "$script:ModuleSourcePath\Resources\Vault\src\PSVault\PSVault.Extension\PSVault.Extension.psm1" -Destination "$script:ArtifactsPath\Resources\Vault\src\PSVault\PSVault.Extension\PSVault.Extension.psm1" -ErrorAction Stop
-    Copy-Item -Path "$script:ModuleSourcePath\Resources\Vault\src\PSScriptAnalyzerSettings.psd1" -Destination "$script:ArtifactsPath\Resources\Vault\src\PSScriptAnalyzerSettings.psd1" -ErrorAction Stop
+    $dirDest = New-Item -Path "$script:ArtifactsPath\$ModuleName.Extension" -ItemType Directory -ErrorAction Stop
+    Copy-Item -Path "$script:ModuleSourcePath\$ModuleName.Extension\*" -Destination $dirDest -Recurse -Force -ErrorAction Stop
     Write-Build Gray '        ...Assets copy complete.'
 }#AssetCopy
 
@@ -443,37 +339,11 @@ Add-BuildTask Build {
     Write-Build White '      Performing Module Build'
 
     Write-Build Gray '        Copying manifest file to Artifacts...'
-    #Copy-Item -Path "$script:ModuleSourcePath)" -Destination $script:ArtifactsPath -Recurse -Exclude 'Archive','Artifacts' -ErrorAction Stop
-    # Copy-Item -Path (Get-Item -Path "$script:ModuleSourcePath\*" -Exclude ('Archive', 'Artifacts')).FullName -Destination $script:ArtifactsPath -Recurse -Force -ErrorAction Stop
-    # Copy-Item -Path (Get-Item -Path "$script:ModuleSourcePath\Resources\*").FullName -Destination $script:ArtifactsPath -Recurse -Force -ErrorAction Stop
     Copy-Item -Path $script:ModuleManifestFile -Destination $script:ArtifactsPath -Recurse -ErrorAction Stop
+    #Copy-Item -Path $script:ModuleSourcePath\bin -Destination $script:ArtifactsPath -Recurse -ErrorAction Stop
     Write-Build Gray '        ...manifest copy complete.'
 
-    Write-Build Gray '        Merging Public and Private functions to one module file...'
-    #$private = "$script:ModuleSourcePath\Private"
-    $scriptContent = [System.Text.StringBuilder]::new()
-    #$powerShellScripts = Get-ChildItem -Path $script:ModuleSourcePath -Filter '*.ps1' -Recurse
-    $powerShellScripts = Get-ChildItem -Path $script:ArtifactsPath -Filter '*.ps1' -Recurse | Where-Object { $_.FullName -notmatch 'Resources' }
-    foreach ($script in $powerShellScripts) {
-        $null = $scriptContent.Append((Get-Content -Path $script.FullName -Raw))
-        $null = $scriptContent.AppendLine('')
-        $null = $scriptContent.AppendLine('')
-    }
-    $scriptContent.ToString() | Out-File -FilePath $script:BuildModuleRootFile -Encoding utf8 -Force
-    Write-Build Gray '        ...Module creation complete.'
-
-    #here we update the parent level docs. If you would prefer not to update them, comment out this section.
-    Write-Build Gray '        Overwriting docs output...'
-    Move-Item "$($script:ArtifactsPath)\docs\*.md" -Destination "..\docs\" -Force
-    Remove-Item "$($script:ArtifactsPath)\docs" -Recurse -Force -ErrorAction Stop
-    Write-Build Gray '        ...Docs output completed.'
-
-    Write-Build Gray '        Cleaning up leftover artifacts...'
-    #cleanup artifacts that are no longer required
-    Remove-Item "$($script:ArtifactsPath)\Imports.ps1" -Force -ErrorAction SilentlyContinue
-    Remove-Item "$($script:ArtifactsPath)\Public" -Recurse -Force -ErrorAction SilentlyContinue
-    Remove-Item "$($script:ArtifactsPath)\Private" -Recurse -Force -ErrorAction SilentlyContinue
-    Write-Build Green '        ...Build Complete!'
+    Write-Build Green '      ...Build Complete!'
 }#Build
 
 #Synopsis: Invokes all Pester Infrastructure Tests in the Tests\Infrastructure folder (if it exists)